'レビュー三回目修正'

+++ tweets/views.py
@@ -1,14 +1,12 @@
-from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.http import JsonResponse
-from django.views.generic import CreateView, ListView, DetailView
-from django.views.generic.edit import DeleteView
-from django.views import View
-from django.urls import reverse_lazy
-from django.contrib.auth.mixins import UserPassesTestMixin
-from accounts.models import User, Friendship
-from .models import Tweet, Favorite
-from .forms import TweetForm
 from django.shortcuts import get_object_or_404
+from django.urls import reverse_lazy
+from django.views import View
+from django.views.generic import CreateView, DetailView, ListView
+from django.views.generic.edit import DeleteView
+from .forms import TweetForm
+from .models import Favorite, Tweet
 
 
 class HomeView(LoginRequiredMixin, ListView):
@@ -24,11 +22,9 @@ class HomeView(LoginRequiredMixin, ListView):
             .filter(user=self.request.user)
             .values_list("tweet")
         )
-        context = {
-            "user_liked_list": user_like_list,
-        }
+        context["user_liked_list"] = user_like_list
 
-        return super().get_context_data(**context)
+        return context
 
 
 class TweetCreateView(LoginRequiredMixin, CreateView):
@@ -50,16 +46,18 @@ class TweetDetailView(LoginRequiredMixin, DetailView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
+
+        tweet = get_object_or_404(Tweet, pk=self.kwargs["pk"])
+
         user_like_list = (
             Favorite.objects.select_related("tweet")
-            .filter(user=self.request.user)
+            .filter(tweet=tweet, user=self.request.user)
             .values_list("tweet")
         )
-        context = {
-            "user_liked_list": user_like_list,
-        }
 
-        return super().get_context_data(**context)
+        context["user_liked_list"] = user_like_list
+
+        return context
 
 
 class TweetDeleteView(UserPassesTestMixin, DeleteView):
@@ -70,43 +68,6 @@ class TweetDeleteView(UserPassesTestMixin, DeleteView):
     def test_func(self):
         tweet_instance = self.get_object()
         return tweet_instance.author == self.request.user
-
-
-class UserProfileView(LoginRequiredMixin, ListView):
-    template_name = "tweets/user_profile.html"
-    model = Tweet
-    ordering = "-created_at"
-
-    def get_queryset(self):
-        author = get_object_or_404(User, username=self.kwargs["username"])
-        return Tweet.objects.select_related("author").filter(author=author)
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        author = get_object_or_404(User, username=self.kwargs["username"])
-        self.followee = Friendship.objects.select_related("followee").filter(
-            follower=author
-        )
-        self.follower = Friendship.objects.select_related("follower").filter(
-            followee=author
-        )
-        is_follow_or_not = Friendship.objects.filter(
-            followee=author, follower=self.request.user
-        ).exists()
-        user_like_list = (
-            Favorite.objects.select_related("tweet")
-            .filter(user=self.request.user)
-            .values_list("tweet")
-        )
-        context = {
-            "profile": author,
-            "follow_or_not": is_follow_or_not,
-            "num_follows": self.followee.count(),
-            "num_followers": self.follower.count(),
-            "user_liked_list": user_like_list,
-        }
-
-        return super().get_context_data(**context)
 
 
 class LikeView(LoginRequiredMixin, View):
@@ -131,7 +92,10 @@ class UnlikeView(LoginRequiredMixin, View):
     def post(self, request, *args, **kwargs):
         tweet = get_object_or_404(Tweet, pk=self.kwargs["pk"])
         user = request.user
-        Favorite.objects.filter(tweet=tweet, user=user).delete()
+        if Favorite.objects.filter(tweet=tweet, user=user).exists():
+            Favorite.objects.filter(tweet=tweet, user=user).delete()
+        else:
+            pass
         num_liked = tweet.favorite_tweet.count()
         context = {
             "num_liked": num_liked,
